Moved TabBrowser into a different namespace
