+++ src/BrowserHost/Tab/TabBrowser.cs
@@ -1,5 +1,7 @@
 ﻿using BrowserHost.CefInfrastructure;
+using BrowserHost.Features.ActionContext;
 using BrowserHost.Features.ActionContext.FileDownloads;
+using BrowserHost.Features.ActionContext.Tabs;
 using BrowserHost.Features.CustomWindowChrome;
 using BrowserHost.Features.DragDrop;
 using BrowserHost.Features.TabPalette.FindText;
@@ -10,7 +12,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Windows;
 
-namespace BrowserHost.Features.ActionContext.Tabs;
+namespace BrowserHost.Tab;
 
 public class TabBrowser : Browser
 {
